+++ CHANGES.rst
@@ -91,6 +91,8 @@ Changelog
 - Better handle overrides between years.
   [cekk]
 
+- Handle possibility to create more overrides with different gates schedule.
+  [cekk]
 
 2.1.5 (2023-11-10)
 ------------------